Move routes to their own respective files
