+++ package.json
@@ -27,7 +27,7 @@
     "typescript": "^4.0.5"
   },
   "scripts": {
-    "start": "node dist/index.js",
+    "start": "node ./dist/index.js",
     "build": "npx tsc -p tsconfig.json"
   },
   "eslintConfig": {
